WIP attempted refactor of original simple skill
