+++ package.json
@@ -22,6 +22,7 @@
     "gulp-rimraf": "^0.2.1",
     "gulp-shell": "^0.5.2",
     "gulp-webpack": "^1.5.0",
-    "gulp-zip": "^3.2.0"
+    "gulp-zip": "^3.2.0",
+    "json-loader": "^0.5.4"
   }
 }
