+++ speechAssets/SampleUtterances.txt
@@ -1,8 +1,1 @@
-HelloWorldIntent say hello
-HelloWorldIntent say hello world
-HelloWorldIntent hello
-HelloWorldIntent say hi
-HelloWorldIntent say hi world
-HelloWorldIntent hi
-HelloWorldIntent how are you
 IntroQuestionIntent {Answer}
