+++ speechAssets/customSlotTypes/LIST_ANSWERS_ROUND_ONE
@@ -1,0 +1,6 @@
+no
+sit
+stop
+down
+fetch
+bad
