--- src/LIST_ANSWERS_ROUND_ONE
@@ -1,6 +1,0 @@
-No
-Sit
-Stop
-Down
-Fetch
-Bad
