+++ src/questionStorage.js
@@ -1,0 +1,9 @@
+import roundQuestions from "./roundQuestions";
+export const CARD_TITLE = "ROUND_ONE";
+export function getWelcomeRoundQuestion (){
+  return roundQuestions[0].question;
+}
+
+export function getRoundQuestions () {
+  return roundQuestions;
+}
