+++ src/roundQuestions.json
@@ -1,0 +1,13 @@
+[
+  {
+    "question": "Name a word that most people yell at their dogs",
+    "answers": [
+      "no",
+      "sit",
+      "stop",
+      "down",
+      "fetch",
+      "bad"
+    ]
+  }
+]
